+++ packages/jazz-nodejs/src/test/startWorker.test.ts
@@ -1,11 +1,9 @@
-import { Peer } from "cojson";
-import { createWebSocketPeer } from "cojson-transport-ws";
 import { createWorkerAccount } from "jazz-run/createWorkerAccount";
 import { startSyncServer } from "jazz-run/startSyncServer";
-import { CoMap, Group, co } from "jazz-tools";
-import { describe, expect, onTestFinished, test, vi } from "vitest";
-import { WebSocket } from "ws";
+import { CoMap, Group, InboxSender, co } from "jazz-tools";
+import { describe, expect, onTestFinished, test } from "vitest";
 import { startWorker } from "../index";
+import { waitFor } from "./utils";
 
 async function setup() {
   const { server, port } = await setupSyncServer();
@@ -34,13 +32,13 @@ async function setupSyncServer(defaultPort = "0") {
 }
 
 async function setupWorker(syncServer: string) {
-  const { accountId, agentSecret } = await createWorkerAccount({
+  const { accountID, agentSecret } = await createWorkerAccount({
     name: "test-worker",
     peer: syncServer,
   });
 
   return startWorker({
-    accountID: accountId,
+    accountID: accountID,
     accountSecret: agentSecret,
     syncServer,
   });
@@ -96,6 +94,42 @@ describe("startWorker integration", () => {
 
     expect(mapOnWorker2?.value).toBe("test");
 
+    await worker2.done();
+  });
+
+  test("reiceves the messages from the inbox", async () => {
+    const worker1 = await setup();
+    const worker2 = await setupWorker(worker1.syncServer);
+
+    const group = Group.create({ owner: worker1.worker });
+    const map = TestMap.create(
+      {
+        value: "Hello!",
+      },
+      { owner: group },
+    );
+
+    worker2.experimental.inbox.subscribe(TestMap, async (value) => {
+      return TestMap.create(
+        {
+          value: value.value + " Responded from the inbox",
+        },
+        { owner: value._owner },
+      );
+    });
+
+    const sender = await InboxSender.load<TestMap, TestMap>(
+      worker2.worker.id,
+      worker1.worker,
+    );
+
+    const resultId = await sender.sendMessage(map);
+
+    const result = await TestMap.load(resultId, worker2.worker, {});
+
+    expect(result?.value).toEqual("Hello! Responded from the inbox");
+
+    await worker1.done();
     await worker2.done();
   });
 });