+++ packages/jazz-tools/src/coValues/account.ts
@@ -14,16 +14,13 @@ import {
 } from "cojson";
 import {
   AnonymousJazzAgent,
-  CoMap,
   type CoValue,
   CoValueBase,
   CoValueClass,
   DeeplyLoaded,
   DepthsIn,
-  Group,
   ID,
   MembersSym,
-  Profile,
   Ref,
   type RefEncoded,
   RefIfCoValue,
@@ -37,6 +34,11 @@ import {
   subscriptionsScopes,
 } from "../internal.js";
 import { coValuesCache } from "../lib/cache.js";
+import { type CoMap } from "./coMap.js";
+import { type Group } from "./group.js";
+import { createInboxRoot } from "./inbox.js";
+import { Profile } from "./profile.js";
+import { RegisteredSchemas } from "./registeredSchemas.js";
 
 /** @category Identity & Permissions */
 export class Account extends CoValueBase implements CoValue {
@@ -59,7 +61,7 @@ export class Account extends CoValueBase implements CoValue {
         optional: false,
       } satisfies RefEncoded<Profile>,
       root: {
-        ref: () => CoMap,
+        ref: () => RegisteredSchemas["CoMap"],
         optional: true,
       } satisfies RefEncoded<CoMap>,
     };
@@ -234,17 +236,25 @@ export class Account extends CoValueBase implements CoValue {
     return this.toJSON();
   }
 
-  migrate(
-    this: Account,
-    creationProps?: { name: string },
-  ): void | Promise<void> {
+  migrate(this: Account, creationProps?: { name: string }) {
     if (creationProps) {
-      const profileGroup = Group.create({ owner: this });
+      const profileGroup = RegisteredSchemas["Group"].create({ owner: this });
       profileGroup.addMember("everyone", "reader");
       this.profile = Profile.create(
         { name: creationProps.name },
         { owner: profileGroup },
       );
+    }
+
+    const node = this._raw.core.node;
+    const profile = node
+      .expectCoValueLoaded(this._raw.get("profile")!)
+      .getCurrentContent() as RawCoMap;
+
+    if (!profile.get("inbox")) {
+      const inboxRoot = createInboxRoot(this);
+      profile.set("inbox", inboxRoot.id);
+      profile.set("inboxInvite", inboxRoot.inviteLink);
     }
   }
 
@@ -388,3 +398,5 @@ export function isControlledAccount(account: Account): account is Account & {
 export type AccountClass<Acc extends Account> = CoValueClass<Acc> & {
   fromNode: (typeof Account)["fromNode"];
 };
+
+RegisteredSchemas["Account"] = Account;