+++ examples/music-player/CHANGELOG.md
@@ -1,5 +1,12 @@
 # jazz-example-musicplayer
 
+## 0.0.40
+
+### Patch Changes
+
+- jazz-react@0.8.44
+- jazz-tools@0.8.44
+
 ## 0.0.39
 
 ### Patch Changes