+++ examples/onboarding/CHANGELOG.md
@@ -1,5 +1,13 @@
 # jazz-example-onboarding
 
+## 0.0.21
+
+### Patch Changes
+
+- jazz-react@0.8.44
+- jazz-tools@0.8.44
+- jazz-browser-media-images@0.8.44
+
 ## 0.0.20
 
 ### Patch Changes