+++ examples/passkey-svelte/CHANGELOG.md
@@ -1,5 +1,11 @@
 # passkey-svelte
 
+## 0.0.9
+
+### Patch Changes
+
+- jazz-svelte@0.8.44
+
 ## 0.0.8
 
 ### Patch Changes