+++ examples/reactions/CHANGELOG.md
@@ -1,5 +1,13 @@
 # reactions
 
+## 0.0.17
+
+### Patch Changes
+
+- jazz-react@0.8.44
+- jazz-tools@0.8.44
+- jazz-browser-media-images@0.8.44
+
 ## 0.0.16
 
 ### Patch Changes